SearchableTextExtender: Don't just decode the Content-Transfer-Encoding,
but also decode the actual content itself with the declared charset.

The Content-Transfer-Encoding (QP or base64) is just an encoding layer
to ensure the message contains only 7-bit characters when sending it over
SMTP. The content itself (if it contains non-ASCII characters) is still
encoded in a character set that maps unicode characters to 8-bit bytes.