Refactor mail view to include all HTML parts.
When a mail has inline attachments, we may have two surrounding HTML parts.
Those parts are complete HTML documents (with <html> tag etc), which broke
the rendering so that only the last part was rendered.

We still consume all HTML parts but extract the body and make them safe
separately, so that we only merge partials.

Inline attachments are still not positioned at the right position but below
in a list of attachments, which is okay.