Drop unnecessary check for regex match in `safe_decode_header`:
For one, the check should have been done using search(), not match() -
we not only want to match the pattern at the beginning of the line,
but anywhere. Once search() would be used, the check becomes entirely
unnecessary, because re.sub() only ever does replacement if there's
one or more matches.