Fix broken <meta /> tags in HTML MIME parts.
text/html MIME parts of a multipart message may include <meta /> tags that
declare a charset that differs from the encoding that is used for the page
template that part is inserted into.

zope.pagetemplate attempts to handle meta tags when inserting markup
with 'structure', by re-encoding the inserted markup and adjusting the
rewriting the meta tag's charset.

If however the <meta /> tag is broken and doesn't match zope.pt's regex,
the re-encoding doesn't take place and TAL / zope.pt chokes on the
resulting mojibake, omitting parts of the document.

This addresses one specific case seen in the wild in a message created
with Apple Mail (2.2104), where the semicolon (;) is missing between the
MIME type and the charset declaration.