Handle more cases of cropped attachment filenames:
The existing condition that was used to try and avoid picking a cropped
filename from the Content-Disposition / Content-Type headers was too
conservative.

It only picked one of the unmangled names when multiple were
present, but we've seen messages in the wild that contain exactly one
unmangled filename in the 'name' param of the Content-Type header (in addition
to a cropped one in the Content-Disposition header).

This change attempts to widen the scope of this fallback to include that case.